+++ src/config.js
@@ -52,9 +52,15 @@ export default Object.assign(
 
     ASSET_TYPES: {
       // Will be updated on init
-      default: '0',
-      kycRequired: '1',
-      security: '2',
+      default: '',
+      kycRequired: '',
+      security: '',
+    },
+
+    POLL_TYPES: {
+      // Will be updated on init
+      restricted: '',
+      unrestricted: '',
     },
 
     CHANGE_ROLE_TASKS: {